+++ crates/snark/src/tests/mod.rs
@@ -1,4 +1,0 @@
-#[cfg(not(target_family = "wasm"))]
-pub mod native;
-#[cfg(target_family = "wasm")]
-pub mod wasm;