+++ crates/snark/src/tests/wasm/mod.rs
@@ -1,1 +1,0 @@
-pub mod test_wasm_witness;