+++ extension/snark/src/tests/mod.rs
@@ -1,2 +1,0 @@
-#[cfg(rings_native)]
-mod test_native;